Refactor AnalyticsViewModel: moved expense logic to AnalyticsService
